rewritten compiler, refactored all classes, full PHP5 interface, exceptions, new plugin architecture
note that the {block} command is currently out of order, and therefore not demonstrated in the test script. it is present, but currently not reliable  - it will be added again when I find a reliable way to implement it.

+++ trunk/class/cache.php
@@ -3,9 +3,9 @@
 /*
 
 OutlineCache
-----------
+------------
 
-Copyright (C) 2008, Rasmus Schultz <http://www.mindplay.dk>
+Copyright (C) 2007-2008, Rasmus Schultz <http://www.mindplay.dk>
 
 Please see "README.txt" for license and usage information.
 
@@ -15,61 +15,54 @@ define("OUTLINE_CACHE_ENGINE", true);
 
 class OutlineCache {
 	
-	var $path;
-	var $time;
-	var $valid = false;
-	var $checked = false;
-	var $buffering = false;
+	protected $path, $time;
 	
-	function OutlineCache($path) {
+	protected $valid = false;
+	protected $checked = false;
+	protected $buffering = false;
+	
+	public function __construct($path) {
 		
-		if (count($path) == 0) trigger_error("Cache: constructor requires one or more arguments", E_USER_ERROR);
+		if (count($path) == 0) throw new OutlineException("Cache: constructor requires one or more arguments");
 		
 		$this->path = array(OUTLINE_CACHE_PATH);
 		
 		$last = count($path);
 		for ($n=0; $n<$last; $n++) {
-			$this->path[] = '/' . OutlineCache::_clean($path[$n]) . ($n == $last-1 ? OUTLINE_CACHE_SUFFIX : '');
+			$this->path[] = '/' . self::_clean($path[$n]) . ($n == $last-1 ? OUTLINE_CACHE_SUFFIX : '');
 		}
 		
 		$this->time = time();
 		
 	}
 	
-	function valid($time_sec = OUTLINE_CACHE_TIME) {
+	public function valid($time_sec = OUTLINE_CACHE_TIME) {
 		
 		if ($this->checked) return $this->valid;
 		
 		$time = @filemtime(implode($this->path));
-		
 		$this->valid = $time ? ($this->time - $time < $time_sec) : false;
-		
 		$this->checked = true;
 		
 		return $this->valid;
 		
 	}
 	
-	function get() {
-		
+	public function get() {
 		if ($this->valid) return implode($this->path);
-		
-		trigger_error("Cache: valid() must be called before get(), and only if valid() returns true");
-		
+		throw new OutlineException("Cache: valid() must be called before get(), and only if valid() returns true");
 	}
 	
-	function capture() {
+	public function capture() {
 		
-		if ($this->buffering) trigger_error("Cache: capture() may only be called once, or after stop() has been called");
+		if ($this->buffering) throw new OutlineException("Cache: capture() may only be called once, or after stop() has been called");
 		
-		$path = '';
-		
-		$last = count($this->path);
+		$path = ''; $last = count($this->path);
 		
 		for ($n=0; $n<$last-1; $n++) {
 			$path .= $this->path[$n];
 			if (!is_dir($path)) {
-				if (!mkdir($path)) trigger_error("Cache: unable to create folder ".$path, E_USER_ERROR);
+				if (!mkdir($path)) throw new OutlineException("Cache: unable to create folder ".$path);
 			}
 		}
 		
@@ -79,12 +72,12 @@ class OutlineCache {
 		
 	}
 	
-	function stop() {
+	public function stop() {
 		
-		if (!$this->buffering) trigger_error("Cache: capture() must be called before stop() can be called");
+		if (!$this->buffering) throw new OutlineException("Cache: capture() must be called before stop() can be called");
 		
 		if (!@file_put_contents(implode($this->path), ob_get_contents()))
-			trigger_error("Cache: unable to write new cache entry ".implode($this->path), E_USER_ERROR);
+			throw new OutlineException("Cache: unable to write new cache entry ".implode($this->path));
 		
 		ob_end_flush();
 		
@@ -92,40 +85,35 @@ class OutlineCache {
 	
 	// --- Static function for flushing cache:
 	
-	function expire() {
+	public static function expire() {
 		
 		$path = OUTLINE_CACHE_PATH;
 		
 		foreach (func_get_args() as $dir)
-			$path .= '/' . OutlineCache::_clean($dir);
+			$path .= '/' . self::_clean($dir);
 		
-		if (is_dir($path)) {
-			OutlineCache::_delete($path);
-		}
+		if (is_dir($path)) self::_delete($path);
 		
-		if (is_file($path.OUTLINE_CACHE_SUFFIX)) {
-			unlink($path.OUTLINE_CACHE_SUFFIX);
-		}
+		if (is_file($path.OUTLINE_CACHE_SUFFIX)) unlink($path.OUTLINE_CACHE_SUFFIX);
 		
 	}
 	
 	// --- Private helper functions:
 	
-	function _clean($dir) {
+	protected static function _clean($dir) {
 		return $dir; /* TO-DO: replace invalid characters */
 	}
 	
-	function _delete($path) {
+	protected static function _delete($path) {
 		
 		$wiped = true;
 		
-		foreach (glob($path . '/*' . OUTLINE_CACHE_SUFFIX) as $file) {
-			if (!unlink($file)) trigger_error("Cache: unable to remove file '$file'", E_USER_NOTICE);
-		}
+		foreach (glob($path . '/*' . OUTLINE_CACHE_SUFFIX) as $file)
+			if (!unlink($file)) trigger_error("Cache: unable to remove file '$file'", E_USER_WARNING);
 		
 		foreach (glob($path . '/*', GLOB_ONLYDIR) as $dir) {
-			$wiped = $wiped && OutlineCache::_delete($dir);
-			if (!rmdir($dir)) trigger_error("Cache: unable to remove dir '$dir'", E_USER_NOTICE);
+			$wiped = $wiped && self::_delete($dir);
+			if (!rmdir($dir)) trigger_error("Cache: unable to remove dir '$dir'", E_USER_WARNING);
 		}
 		
 		return $wiped;
